[audio][drivers][as370] Add support for concurrent dHub DMAs
We use 2 concurrent DMAs to have most of the time 2 DMAs in flight.
Without concurrent DMAs a new DMA needs to be re-enabled within
the time it takes an MTU to be drained (e.g. 333 usecs), since the
interrupt is triggered once the last MTU is transferred.
With concurrent DMAs a new DMA can be safely delayed for up
to the time it takes to drain a complete DMA (e.g. 21.3 msecs),
i.e. MTU size x dma_mtus.

Also allow the dHub thread to shutdown on a stopped thread.

Bug: 107645
Test: fx test -ov syn-dhub-test
Test: Added a 19 msecs delay in the interrupt processing guarantying
the next new DMA to not be enabled in time. Without this CL, this
guarantees audio glitches for audio heard on output and recorded for
input.

Change-Id: Iad80c67017b535d11dd932e8393c4e32c9382980
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/782127
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Jack Brennen <jbrennen@google.com>