[superproject] Roll third_party/pigweed pw_build: Add pw-wrap-ninja helper
This commit adds a standalone Python script that wraps Ninja to provide
a more comprehensive display of the progress of the build, along with a
brief summary of the tool to the pw_build docs.

The tool works by parsing Ninja's output to build a representation of
the build status, and then outputting a richer TUI with the list of
running targets and how long they're taking. This allows one to, at a
glance, determine what targets are blocking the rest of the build or
taking an unusually long time to complete.

Original-Bug: 263976434
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/125551
Original-Revision: 21a65831d7e63b8ef73506c69627fefb6a7c6634
GitOrigin-RevId: e0c1279be86503d709c473dad8d28f06cbe51ca8
Roller-URL: https://ci.chromium.org/b/8793387337694284065
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I9a99385e6cceceb31a2fc311fb04f11fadfa92fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784424