[starnix] Refactor binder reference counting
The binder reference counting need to be able to notify the owner of a
reference whenever its weak or strong reference count reach 0, or
increase from 0.

The previous implementation only had reference count at the handle
level, and was relying on Rust Arc to keep the referenced object alive.
This prevented to send the correct information to the client.

This CL updates the object to have their own reference count. Moreover,
it handled sending weak count back to the owner, and handles receiving
ack from the owner for creation of weak and strong reference to an
object.

Change-Id: I0bdea9191052c29ec9f9966656f4bec36cae990c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/780402
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>