[identity] Update AccountHander README
This is something I've been wanting to do for a while because AH is now
our largest component and its got the most people working in it while
the documentation was woefully out of date.

Happy to make changes to the text to make it clearer and fix any errors.
Going forwards, this is about the level of detail I'd like us to
maintain for the modules so in many cases we can have a discussion about
the implementation of an epic based on the changes it wants to make to
the scope of the modules in this readme.

Change-Id: Id1e5fdb4ebf43212c6588edae8f918f861baa9f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/783508
Reviewed-by: James Buckland <jbuckland@google.com>
Reviewed-by: Angad Pal Singh <apsbhatia@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>