[emu] Add support for PBv2 with multiple VDs
This CL adds conversion code to the `ffx emu` plugin for an updated form
of PBv2 with virtual devices linked to the PBM file. This is based on
https://fxrev.dev/758802, but uses the kernel file from system_a and the
template file linked to the selected virtual device.

The virtual devices are stored in a new VirtualDeviceManifest, which is
a separate file linked to the PBM through the virtual_devices_path
field. The manifest type holds a map of names to paths to virtual
device files, and an optional recommended default device name.

Bug: 116056
Bug: 113144
Bug: 90948

Change-Id: I9301e63cf8ca269b2cc1f5183850244bd1326d5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/774004
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>