[zircon] Remove bad assertion in VMAR Enumerator.
This CL removes a DEBUG_ASSERT that was in VmAddressRegionEnumerator,
that would check whether the next value in the enumerator is a mapping,
disallowing vmars.

From the bug comments:

> `itr_` points to the item after the one that was just yielded by
> `next()`, and that could either be a mapping or a vmar.

This debug assert could be triggered by issuing an operation on a vmar
that has another vmar at the end of the range. The CL adds a test that
exercises this scenario.

BUG: 118226
Change-Id: I5347ef3ef52afcbb947f7033fb4516658300058c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785564
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>