[rust-3p] Update notify to 5.0.0
This patch updates notify from 5.0.0-pre.10 to 5.0.0. It also imports
the kqueue 1.0.7, and kqueue-sys 1.0.3, which was extracted from notify
(approved in OSRB http://fxbug.dev/116738).

As part of this, this disables notify from internally using
crossbeam-channel because of
https://github.com/notify-rs/notify/issues/380, where crossbeam will
park the current thread if there's nothing on the channel. This can
cause a deadlock when running in an async executor. std::sync::mpsc
doesn't do this, so this should avoid any risk of deadlocks.

Finally, this forks the notify crate because the 5.0.0 accidentally left
out the license file. This will be fixed with
https://github.com/notify-rs/notify/issues/460.

Bug: 116738
Change-Id: Ic058a2d6ed71cd677fc8f7acb5083405a0bf4884
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/776923
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Reviewed-by: Dan Johnson <computerdruid@google.com>
Reviewed-by: Viktar Donich <viktard@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>