[moniker] Remove `rep` field from ChildMoniker
ChildMoniker stores the display representation of the moniker
as a string, in addition to the separate strings for the
collection and name.

This uses memory unnecessarily. Construct the display
representation when it is needed using the `ToString` or
`Display` trait.

Change-Id: I9d897c154270fe5944517463d5a8d64b9fcdb895
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785602
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>