[settings] Warn on missing registrant dependency
Rather than silently skipping the registration of a `Registrant` that is
missing a dpendency we now log that the registration is being skipped.
The name of the registrant is now tracked so that it can be reported
when there is an error, to support that some cleanup was done as well.

Tests: fx setui_service_tests
Fixed: 96251
Change-Id: I49aee28998586ee11c5dbeea5668a0c973937dac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784442
Reviewed-by: Paul Faria <paulfaria@google.com>
Commit-Queue: Eric Rahm <erahm@google.com>