[netemul] Keep strong ref to realm from Interface
TestInterface had to eagerly connect to a number of services in its
realm so fulfill its helper functions.

By having access to its containing realm we can connect to the services
when needed.

Ideally, TestInterface would keep a borrow of the realm so that we can
encode the lifetimes, but that makes helper functions that returns
TestRealm + TestInterface impossible because that would be a
self-referential return.

Prior to this change, destroying a realm before the TestInterface would
cause most of the helpers to fail regardless, because the realm
containing the netstack backing the services would be tearing down.

With this change, TestInterface clones the ManagedRealm proxy, keeping
effectively a clone of the TestRealm for itself, which allows it to
connect to services on demand. An unexplored benefit is that helper
functions might drop TestRealms if they only care about a single
interface installed in that realm.

Change-Id: I31872764485992588a7d9a0a7d5d0520c6d63609
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/784303
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>