[media][codec] Implement roundtrip test for CVSD.
Implement test suite for doing roundtrip testing for codecs. It performs
the roundtrip test by calculating the target RMSE from golden files and
running the input through the encoder and the decoder. It compares the
resulting RMSE from the codec-under-test against the target RMSE to
ensure that the RMSE is within the limit. It also performs one-way
test using just the golden encoded file and the decoder to ensure that
the codec works as intended.

Change-Id: I548453d46e4678c964c639e34adbb9637725ecee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759303
Commit-Queue: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>