[moniker] Make &str -> ChildMoniker fallible
We implemented `From<&str>` for ChildMoniker and unwrapped the
result implicitly. This was mostly used in tests, but is a
dangerous implementation because it is also available for
production use.

The Rust Language also specifies that the From trait should
only be used for infallible conversions.

This CL implemented `TryFrom<&str>` for ChildMoniker and
changes all tests to unwrap explicitly.

Bug: 118649
Change-Id: Id60a320f7aee30239bde1f2ac6d35f77c24a81a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786022
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>