[fdf] Replace wlan-softmac's DdkServiceConnect
Make the device `wlan-softmac-device` not implement `DdkServiceConnect`.
Replace it with a new outgoing directory that serves the WlanSoftmac
service and offer the service when creating `wlan-softmac-device`.

This is to avoid `wlan-softmac-device` crashing when run in the DFv2
compat shim. This is because the compat shim tries to serve the FIDL
protocol `RuntimeConnect` on its outgoing directory when the outgoing
directory is already serving the protocol. The compat shim tries to
serve the protocol because `wlan-softmac-device` is service-connectable.
The outgoing directory is already serving the protocol because the
`iwlwifi-wlanphy` device is also service-connectable and the compat shim
started serving the protocol when that device was initialized.
`wlan-softmac-device` and `iwlwifi-wlanphy` share the same outgoing
directory because they share the same driver and the compat shim uses
the outgoing directory associated with the driver when serving the
`RuntimeConnect` FIDL protocol.

This is tested in
`src/connectivity/wlan/drivers/third_party/intel/iwlwifi/test/wlan-softmac-device-test.cc`
where when a `wlan-softmac-device` device is created, the tests will
attempt to establish a connection to the WlanSoftmac protocol found in
the device's outgoing directory.

Bug: 115001
Change-Id: I3aa51456b57c62833d22b2722fecb75162e3847a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761563
API-Review: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>