[settings] Allow missing camera on restore
Rather than failing to bring up the service when the camera is missing,
allow the service to pass restore. In this state the camera status
should show that it's in an error state. Allow clients to attempt to
update the state of the camera, since it could reconnect and bring the
camera out of the error state. If the camera is still stuck, the state
will still show as an error in future requests.

Bug: 117854
Test: fx test -o setui_service_tests
Change-Id: I3c0e0ef8bfcc0e4f54afa4eed079509947012e4e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785510
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>