[rust_crates] Update url to 2.3.1
This updates url from 2.1.0 to 2.3.1. In addition, it updates two
dependencies from the same github repo:

* Imports idna 0.3.0 (in addition to 0.1.5 and 0.2.3)
* Imports form_urlencoded 1.1.0 (approved by OSRB in
  fxbug.dev/114336)

Adds no new unsafe blocks.

The motivating reason for the upgrade is to both improve Fuchsia's
handling of URLs to be spec-compliant, including the ability to use
`url.make_relative()` to improve Fuchsia's input and output (string
parsing and display) of relative URLs.

Note that the `url` crate upgrades are meant to improve URL handling and
spec compliance over previous versions, but this has resulted in some
behavior changes in the `url` crate requiring a few changes to how Fuchsia
parses package and component URLs. Notably:

* `url.path()` will sometimes return an empty string in situations
  that the previous Fuchsia crate returned a "/" by default.
  This appears to have changed in version 2.1.1 (see
  https://github.com/servo/rust-url/issues/773).
* `Url::parse()` used to ignore null characters included in the input
  string (at least in fragments, but perhaps elsewhere). Null
  characters are now percent-encoded in the parse URL.
* `base_url.join(relative_path)` can return a different result,
  depending on the base URL and path. For instance, base `scheme:///`
  combined with `///path` results in `scheme:///path` and removing
  the base (via `base_url.make_relative(url)`) returns simply
  `path`.

Adjustments have been made in Fuchsia libraries. In general, the
affected areas found so far were tests meant to validate extreme cases
not expected to be encountered.

If there are no other issues, it may still be possible to complete this
upgrade, with some mindful attention.

Bug: 114336

Change-Id: I3be46831bb1c2c24d3718f17d51b10b4784b0258
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/779685
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>