[emu] Add a Large virtual device
The attached bug requests a way for Chrome to run with additional
storage capacity in the Fuchsia emulator. Storage is defined in the
virtual device specification.

This CL adds a new virtual device spec to the GN build,
"${device_name}-large", which increases storage capacity of the FVM to
10GB, as per Chrome's request.

Bug: 116366
Change-Id: I3046ee28afcf7d65ff7ae55e9a4833778cca75da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786142
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Bryant Chandler <bryantchandler@google.com>
Reviewed-by: Shai Barack <shayba@google.com>