[ffx][driver] Add `--fail-on-missing` flag to `ffx driver list-devices`
This makes it easier to use `ffx driver list-devices` on shell scripts
to detect the presence of a certain device and fail otherwise.

Preferably we would always silently return a non-zero exit code when no
matching drivers are found, but FFX will print an error message and a
backtrace if a subcommand returns an error, which is not desirable as
the default for this case.

Bug: 113465
Change-Id: Ia390769b7d4138250fe768f4f928b282038c1bcf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/781346
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Ian Fisher <iafisher@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>