[fdf] Update DFv1 composite device workflow
Make changes to the DFv1 composite device workflow
so that a node representing the composite is
only added to the node topology when there's a
driver bound to it.

CompositeDevice now stores a |driver_| object, which
is set by three methods:
1) If CompositeDevice is created by
   CreateFromDriverIndex().
2) CompositeDevice::TryAssemble() - if this function
   is called and |driver_| is nullopt, then
   CompositeDevice will try to find a matching driver
   to set |driver_|.
3) BindDriverManager::BindAllDevices() - when this
   function is called, it'll try to bind a matching
   driver to all devices and CompositeDevice objects.
   This is so if a matching driver is loaded after
   CompositeDevice::TryAssemble() was called, the
   driver can now bind to actual device.

This CL also makes changes to device_enumeration and
src/devices/driver_manager tests.

For device_enumeration Sherlock test, aml-hevc-enc is
removed since the driver that binds to it is
excluded in bringup.

In driver_manager tests, MultipleDeviceTestCase is
modified so that creating the Coordinator config is
overridable by subclasses. Some common code are also
refactored in CompositeDeviceTests.

Bug:118642
Tests: Update driver_manager tests. Added test to
ensure the device is only created when it has a
matching driver

Change-Id: I849f763bf39e6c5ef3d0b2e5b3ab1f8b6167a3be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785483
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>