[cargo-gnaw] Add support for building unit test targets.
This CL adds support for generating executable targets for each library
and binary with tests. We only generate tests for targets that
explicitly enable it in the root Cargo.toml.

This CL does not attempt to deal with hooking these tests up to the rest
of the build.

Change-Id: Ifdc16a4184a862ad26ffd23b57eecd021b6c716f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/765965
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>