[elfldltl] Disable hwasan on static globals in memory test
Hwasan reports a tag mismatch for globals accessed via the memory API
since the tag for the image start will almost never be the same as the
tag for an arbitrary global within that image. The test is WAI, but just
isn't compatible with hwasan.

Fixed: 118713
Change-Id: Ic74299bb3201892246751d4db15fe1be694816af
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786555
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>