[kernel][vm] Remove incorrect can_borrow_locked assert
The value of is_any_borrowing_enabled() can change between when it is
used to evaluate the test condition and the assert. Remove the assert
and replace it with a check for debug_is_user_pager_backed() which is
the only non-redundant portion of the assert anyway.

Fixed: 118821
Run-All-Tests: true
Change-Id: Id8058a570501d4dcd0352dc7e086b664f8727d5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786428
Fuchsia-Auto-Submit: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>