[netstack3] Factor out multi-nic network setup
Factor out a helper that sets up hosts in a star topology
where the multi-nic host is connected to each peer through a separate
network, and none of the peers are connected to each other. This is
useful for testing that sockets can be bound to devices explicitly or
implicitly via zoned addressing.

Change-Id: I52675962ac09adc369cf262b7a35a886f1c9be4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/785563
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>