[fidlgen_rust] Derive Copy/Clone based on resourceness
This CL makes it so that resource types in Rust never derive Copy and
Clone, even when they have no handles. This was the intention all along:
marking a type as `resource` is an up front decision that means adding a
handle later will not be source breaking.

Note that the presence of handles is still used for deriving zerocopy
traits, since these are meant to be an encoding implementation detail.
(Technically this could cause source breakage if someone used them, but
that already happens with vectors and Copy or floats and Eq/Ord/Hash).

Added an empty resource struct, table, and union in
handles_in_types.test.fidl to demonstrate.

Also refactored the derives algorithm to make it simpler.

Fixed: 118784
Change-Id: I5e91b2327578913f16162d75ed06e10f79b59495
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786438
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Mitchell Kember <mkember@google.com>
Reviewed-by: Zachary Stewart <zstewart@google.com>