[fdf] Reland Update DFv1 composite device workflow
This is a reland of
commit 91cfed6078017bffac23439c71232a7633725599.

To prevent the CompositeDeviceTests from hanging,
we now shutdown the FakeDriverIndex's thread.

In addition, error handling in BindAllDevices() is
adjusted so that it'll continue to match and bind
the other composites even when it encounters an
error.

Multiply: fuchsia-pkg://fuchsia.com/driver-manager-tests#meta/driver-manager-test.cm : 100
Change-Id: Ie9b628af5c9673cda480d624fbcc3c836ba1621e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/786848
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>