[docgen] Support templates.
Add support for templates and template specializations now output by
clang-doc.

Support the new "QualName" member on References which allows removing
some custom formatting code to derive this from the path.

Clean up the definition of "Type" to match ClangDoc (now it has a
Reference member), remove the "ReturnType" structure since this was
based on a misunderstanding of clangdoc's output. Add some helper
functions to construct.

Remove the "type" syntax highlighting for the name in record
definitions. It looks weird.

Re-enable the end-to-end test.

Change-Id: I36620e9ed31b5d50167d9bc07847657195d70e6e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/770388
Reviewed-by: Dave Smith <smithdave@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>