[ffx][fho] Implement searching for sdk ffx commands
Use the sdk metadata to discover ffx_tool elements and run them as ffx
subtools when found.

As part of this, some functions in the ffx_command library become async
because we need to at some point load config to find the sdk.

Fixed: 116732
Change-Id: I38584ef615ac09a71565373cd3fb95b5bc037771
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/778886
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>